+++ experimental/CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :rocket: (Enhancement)
 
+* feat(sdk-logs): added colors option to ConsoleLogRecordExporter
 * refactor(instr-http): use exported strings for semconv. [#4573](https://github.com/open-telemetry/opentelemetry-js/pull/4573/) @JamieDanielson
 * feat(sdk-node): add `HostDetector` as default resource detector